create dataformat module - add value format and clean operations to module to remove tangentially related methods from parser module
